+++ src/CommunityPatchLauncher/Commands/Os/StartSettlersConfigCommand.cs
@@ -1,0 +1,30 @@
+﻿using System.Collections.Generic;
+using System.IO;
+using CommunityPatchLauncherFramework.Settings.Manager;
+
+namespace CommunityPatchLauncher.Commands.Os
+{
+    /// <summary>
+    /// starts the settlers 4 he configuration file
+    /// </summary>
+    class StartSettlersConfigCommand : StartProgramCommand
+    {
+        /// <summary>
+        /// The settings manager to use
+        /// </summary>
+        private SettingManager settingManager;
+
+        /// <summary>
+        /// Starts the settlers 4 he configuration file
+        /// </summary>
+        /// <param name="settingManager"></param>
+        public StartSettlersConfigCommand(SettingManager settingManager)
+        {
+            string gameFolder = settingManager.GetValue<string>("GameFolder");
+            string configPath = gameFolder + Properties.Settings.Default.ConfigBaseFolder + "Settlers4Config.exe";
+
+            this.settingManager = settingManager;
+            programPath = configPath;
+        }
+    }
+}